requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
youtube-search-python==1.6.6
pydantic==2.7.1
